:rocket: move timer logic to blackhole component #443
move timer logic to blackhole component #443
