+++ backend/src/blackhole/blackhole.module.ts
@@ -1,25 +1,17 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { BlackholeService } from './blackhole.service';
-import { IBlackholeRepository } from './repository/blackhole.repository';
-import { RawqueryBlackholeRepository } from './repository/rawquery.blackhole.repository';
 import { AuthModule } from 'src/auth/auth.module';
 import { CabinetModule } from 'src/cabinet/cabinet.module';
 import { LentModule } from 'src/lent/lent.module';
 import { UserModule } from 'src/user/user.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-
-const repo = {
-  provide: IBlackholeRepository,
-  useClass: RawqueryBlackholeRepository,
-};
+import { BlackholeTools } from './blackhole.component';
 
 @Module({
   imports: [
     AuthModule, CabinetModule, HttpModule, UserModule, LentModule,
-    // EventEmitterModule,
   ],
-  providers: [BlackholeService, repo],
-  // exports: [BlackholeService]
+  providers: [BlackholeService, BlackholeTools],
 })
 export class BlackholeModule {}
